1. Each forest maintain a variable order. That order is the same as the one maintained in the domain, except during reordering.
2. Modify the cost functions used in Lowest Cost reordering heuristic: number of nodes at upper level * bound of variable at upper level * bound of variable at lower level + number of nodes at lower level

3. Modify dining_phils.cc, kanban.cc, test_union_mdd.cc and test_union_mxd.cc for reordering test.  Create reorder.h to provide common functions used in reordering test.


git-svn-id: svn://svn.code.sf.net/p/meddly/code/branches/cjiang@570 14966d26-f10a-4722-9547-d5dcc201c352

+++ src/heap.h
@@ -15,7 +15,7 @@ typedef struct Inversion
 {
 	// An inversion is formed by the variables at level and level+1
 	int level;
-	int cost;
+	long cost;
 } Inversion;
 
 class InversionHeap
@@ -105,7 +105,7 @@ public:
 	}
 
 	// If the level is in the heap, update the cost
-	void push(int level, int cost)
+	void push(int level, long cost)
 	{
 		if(!is_in_heap(level)){
 			_indices[level]=_heap.size();
@@ -113,7 +113,7 @@ public:
 			percolate_up(level);
 		}
 		else {
-			int old_cost=_heap[_indices[level]].cost;
+			long old_cost=_heap[_indices[level]].cost;
 			_heap[_indices[level]].cost=cost;
 			if(old_cost<cost) {
 				percolate_down(level);
@@ -126,6 +126,7 @@ public:
 
 	void pop()
 	{
+//		printf("%d\n", _heap[0].cost);
 		int top=_heap[0].level;
 		_heap[0]=_heap.back();
 		_indices[_heap[0].level]=0;
